Locate `TestVersionContext` in its own source file
